move loadouts part of job tab into separate class otherwise itll bother me
